requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
boto3==1.34.144
botocore==1.34.144
celery==5.3.6
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==4.4.2
Django==4.2.16
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
moviepy==1.0.3
numpy==2.0.2
packaging==24.1
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
psycopg2-binary==2.9.6
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
